each actor is a singleton. moved some files
